Apply changes from upstream
https://github.com/qt/qtbase/commit/13b3545e833f6175f686c9776e1510db3f3f11eb
https://github.com/qt/qtbase/commit/51d40d7e9bdfc63c5109aef5b732aa2ba10f985a
https://github.com/qt/qtbase/commit/515e802ae20c045e5c47b400ee6ef6e92349c978

+++ src/gui/math3d/qgenericmatrix.cpp
@@ -272,7 +272,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix2x2 type defines a convenient instantiation of the
-    QGenericMatrix template for 2 columns, 2 rows, and qreal as
+    QGenericMatrix template for 2 columns, 2 rows, and float as
     the element type.
 */
 
@@ -281,7 +281,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix2x3 type defines a convenient instantiation of the
-    QGenericMatrix template for 2 columns, 3 rows, and qreal as
+    QGenericMatrix template for 2 columns, 3 rows, and float as
     the element type.
 */
 
@@ -290,7 +290,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix2x4 type defines a convenient instantiation of the
-    QGenericMatrix template for 2 columns, 4 rows, and qreal as
+    QGenericMatrix template for 2 columns, 4 rows, and float as
     the element type.
 */
 
@@ -299,7 +299,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix3x2 type defines a convenient instantiation of the
-    QGenericMatrix template for 3 columns, 2 rows, and qreal as
+    QGenericMatrix template for 3 columns, 2 rows, and float as
     the element type.
 */
 
@@ -308,7 +308,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix3x3 type defines a convenient instantiation of the
-    QGenericMatrix template for 3 columns, 3 rows, and qreal as
+    QGenericMatrix template for 3 columns, 3 rows, and float as
     the element type.
 */
 
@@ -317,7 +317,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix3x4 type defines a convenient instantiation of the
-    QGenericMatrix template for 3 columns, 4 rows, and qreal as
+    QGenericMatrix template for 3 columns, 4 rows, and float as
     the element type.
 */
 
@@ -326,7 +326,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix4x2 type defines a convenient instantiation of the
-    QGenericMatrix template for 4 columns, 2 rows, and qreal as
+    QGenericMatrix template for 4 columns, 2 rows, and float as
     the element type.
 */
 
@@ -335,7 +335,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix4x3 type defines a convenient instantiation of the
-    QGenericMatrix template for 4 columns, 3 rows, and qreal as
+    QGenericMatrix template for 4 columns, 3 rows, and float as
     the element type.
 */
